feat: create inventory Athena table in stack
