+++ cdk/stack.py
@@ -1,6 +1,7 @@
-from typing import Any
+from typing import Any, cast
 
-from aws_cdk import Duration, Stack
+from aws_cdk import Duration, RemovalPolicy, Stack
+from aws_cdk import aws_glue as glue
 from aws_cdk import aws_iam as iam
 from aws_cdk import aws_lambda as lambda_
 from aws_cdk import aws_lambda_event_sources as sources
@@ -17,6 +18,7 @@ class HlsLpdaacReconciliationStack(Stack):
         scope: Construct,
         construct_id: str,
         *,
+        hls_inventory_reports_id: str,
         hls_inventory_reports_bucket: str,
         hls_forward_bucket: str,
         hls_historical_bucket: str,
@@ -44,6 +46,13 @@ class HlsLpdaacReconciliationStack(Stack):
         # Generate HLS product inventory report
         # ----------------------------------------------------------------------
 
+        self.inventory_table = self.make_inventory_table(
+            location=f"s3://{hls_inventory_reports_bucket}/{hls_forward_bucket}/{hls_inventory_reports_id}/hive",
+        )
+        inventory_table_name = cast(
+            glue.CfnTable.TableInputProperty, self.inventory_table.table_input
+        ).name
+
         # Bucket where HLS inventory reports are written.
         inventory_reports_bucket = s3.Bucket.from_bucket_name(
             self, "HlsInventoryReportsBucket", hls_inventory_reports_bucket
@@ -63,6 +72,7 @@ class HlsLpdaacReconciliationStack(Stack):
                 "QUERY_OUTPUT_PREFIX": f"s3://{hls_inventory_reports_bucket}/queries",
                 "REPORT_OUTPUT_PREFIX": f"s3://{hls_inventory_reports_bucket}/reconciliation_reports",
                 "HLS_PRODUCT_VERSION": "2.0",
+                "INVENTORY_TABLE_NAME": inventory_table_name,
             },
         )
         inventory_reports_bucket.grant_read_write(inventory_report_lambda)
@@ -144,3 +154,55 @@ class HlsLpdaacReconciliationStack(Stack):
         s3.Bucket.from_bucket_name(
             self, "LpdaacReconciliationReports", lpdaac_reconciliation_reports_bucket
         ).grant_read(lpdaac_response_lambda)
+
+    def make_inventory_table(self, *, location: str) -> glue.CfnTable:
+        from aws_cdk.aws_glue import CfnTable
+
+        # See https://docs.aws.amazon.com/AmazonS3/latest/userguide/storage-inventory-athena-query.html
+        table = CfnTable(
+            self,
+            "hls_inventory",
+            catalog_id=self.account,
+            database_name="default",
+            table_input=CfnTable.TableInputProperty(
+                # Table names cannot contain dashes, so replace them with underscores
+                name=f"hls_inventory_{self.stack_name.replace('-', '_')}",
+                owner="hadoop",
+                table_type="EXTERNAL_TABLE",
+                parameters={
+                    "projection.enabled": "true",
+                    "projection.dt.type": "date",
+                    "projection.dt.format": "yyyy-MM-dd",
+                    "projection.dt.range": "2025-06-03,NOW",
+                    "projection.dt.interval": "1",
+                    "projection.dt.interval.unit": "DAYS",
+                },
+                partition_keys=[
+                    CfnTable.ColumnProperty(name="dt", type="string"),
+                ],
+                storage_descriptor=CfnTable.StorageDescriptorProperty(
+                    columns=[
+                        CfnTable.ColumnProperty(name="bucket", type="string"),
+                        CfnTable.ColumnProperty(name="key", type="string"),
+                        CfnTable.ColumnProperty(name="size", type="bigint"),
+                        CfnTable.ColumnProperty(
+                            name="last_modified_date", type="timestamp"
+                        ),
+                    ],
+                    input_format="org.apache.hadoop.hive.ql.io.SymlinkTextInputFormat",
+                    location=location,
+                    number_of_buckets=-1,
+                    output_format="org.apache.hadoop.hive.ql.io.HiveIgnoreKeyTextOutputFormat",
+                    serde_info=CfnTable.SerdeInfoProperty(
+                        serialization_library="org.apache.hadoop.hive.ql.io.parquet.serde.ParquetHiveSerDe",
+                        parameters={
+                            "serialization.format": "1",
+                        },
+                    ),
+                ),
+            ),
+        )
+
+        table.apply_removal_policy(RemovalPolicy.DESTROY)
+
+        return table
