+++ cdk/stack_it.py
@@ -1,7 +1,6 @@
-from typing import Any, cast
+from typing import Any
 
 from aws_cdk import CfnOutput, RemovalPolicy, Stack
-from aws_cdk import aws_glue as glue
 from aws_cdk import aws_iam as iam
 from aws_cdk import aws_s3 as s3
 from aws_cdk import aws_sns as sns
@@ -32,10 +31,6 @@ class HlsLpdaacReconciliationStackIT(Stack):
                 )
             )
 
-        self.inventory_table = self.make_inventory_table(
-            location="s3://impact-hls-inventories/hls-global-v2-forward/HLS_data_products_parquet/hive",
-        )
-
         self.hls_inventory_reports_bucket = self.make_bucket("HlsInventoryReports")
         self.hls_forward_bucket = self.make_bucket("HlsForward")
         self.hls_historical_bucket = self.make_bucket("HlsHistorical")
@@ -51,16 +46,29 @@ class HlsLpdaacReconciliationStackIT(Stack):
             subs.SqsSubscription(lpdaac_request_queue)
         )
 
-        # Set outputs for use within integration tests
-
-        table_input = cast(
-            glue.CfnTable.TableInputProperty, self.inventory_table.table_input
+        self.hls_forward_bucket_inventory_id = "hls_v2_parquet"
+        self.hls_forward_bucket.add_inventory(
+            enabled=True,
+            destination=s3.InventoryDestination(
+                bucket=s3.Bucket.from_bucket_name(
+                    self,
+                    "HlsInventoryReportsBucket",
+                    self.hls_inventory_reports_bucket.bucket_name,
+                ),
+                prefix=None,
+            ),
+            inventory_id=self.hls_forward_bucket_inventory_id,
+            format=s3.InventoryFormat.PARQUET,
+            frequency=s3.InventoryFrequency.DAILY,
+            objects_prefix=None,
+            optional_fields=["Size", "LastModifiedDate"],
         )
 
+        # Set outputs for use within integration tests
         CfnOutput(
             self,
-            "HlsInventoryTable",
-            value=str(table_input.name),
+            "HlsInventoryReportId",
+            value=self.hls_forward_bucket_inventory_id,
         )
         CfnOutput(
             self,
@@ -101,55 +109,3 @@ class HlsLpdaacReconciliationStackIT(Stack):
             auto_delete_objects=True,
             removal_policy=RemovalPolicy.DESTROY,
         )
-
-    def make_inventory_table(self, *, location: str) -> glue.CfnTable:
-        from aws_cdk.aws_glue import CfnTable
-
-        # See https://docs.aws.amazon.com/AmazonS3/latest/userguide/storage-inventory-athena-query.html
-        table = CfnTable(
-            self,
-            "hls_inventory",
-            catalog_id=self.account,
-            database_name="default",
-            table_input=CfnTable.TableInputProperty(
-                # Table names cannot contain dashes, so replace them with underscores
-                name=f"hls_inventory_{self.stack_name.replace('-', '_')}",
-                owner="hadoop",
-                table_type="EXTERNAL_TABLE",
-                parameters={
-                    "projection.enabled": "true",
-                    "projection.dt.type": "date",
-                    "projection.dt.format": "yyyy-MM-dd",
-                    "projection.dt.range": "2025-06-03,NOW",
-                    "projection.dt.interval": "1",
-                    "projection.dt.interval.unit": "DAYS",
-                },
-                partition_keys=[
-                    CfnTable.ColumnProperty(name="dt", type="string"),
-                ],
-                storage_descriptor=CfnTable.StorageDescriptorProperty(
-                    columns=[
-                        CfnTable.ColumnProperty(name="bucket", type="string"),
-                        CfnTable.ColumnProperty(name="key", type="string"),
-                        CfnTable.ColumnProperty(name="size", type="bigint"),
-                        CfnTable.ColumnProperty(
-                            name="last_modified_date", type="timestamp"
-                        ),
-                    ],
-                    input_format="org.apache.hadoop.hive.ql.io.SymlinkTextInputFormat",
-                    location=location,
-                    number_of_buckets=-1,
-                    output_format="org.apache.hadoop.hive.ql.io.HiveIgnoreKeyTextOutputFormat",
-                    serde_info=CfnTable.SerdeInfoProperty(
-                        serialization_library="org.apache.hadoop.hive.ql.io.parquet.serde.ParquetHiveSerDe",
-                        parameters={
-                            "serialization.format": "1",
-                        },
-                    ),
-                ),
-            ),
-        )
-
-        table.apply_removal_policy(RemovalPolicy.DESTROY)
-
-        return table
